+++ CHANGELOG.md
@@ -3,7 +3,16 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
-## Version 0.7.8 - Unreleased
+## Version 0.8.0 - Unreleased
+
+### Removed
+* Drop Python 3.6 and 3.7 support
+
+### Changed
+* Speed improvements to delayed image loading
+
+
+## Version 0.7.8 - Released 2024-03-19
 
 ### Added